manager page

+++ src/main/java/kr/cse/gt/model/BasicSensorInfo.java
@@ -2,17 +2,18 @@ package kr.cse.gt.model;
 
 public class BasicSensorInfo {
 	private Double temperature;
-	private Double humidity;
+	private Integer  humidity;
+	
 	public Double getTemperature() {
 		return temperature;
 	}
 	public void setTemperature(Double temperature) {
 		this.temperature = temperature;
 	}
-	public Double getHumidity() {
+	public Integer getHumidity() {
 		return humidity;
 	}
-	public void setHumidity(Double humidity) {
+	public void setHumidity(Integer humidity) {
 		this.humidity = humidity;
 	}
 }
